fix int/long mismatch

+++ examples/hadoop_word_count/src/WordCount.java
@@ -130,7 +130,7 @@ public class WordCount extends Configured implements Tool
             context.write(outputKey, Collections.singletonList(getMutation(word, sum)));
         }
 
-        private static Mutation getMutation(Text word, long sum)
+        private static Mutation getMutation(Text word, int sum)
         {
             Column c = new Column();
             c.setName(Arrays.copyOf(word.getBytes(), word.getLength()));
